pyproject.toml
[tool.poetry]
name = "auctionpie"
version = "0.1.0"
description = ""
license = "BSD"
authors = ["Sergey M <magaster@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
black = "^20.8b1"
Django = "^3.1.5"
graphene-django = "^2.15.0"
django-graphql-auth = "^0.3.15"
PyJWT = "1.7.1"
flake8 = "^3.8.4"
mypy = "^0.790"
pydantic = "^1.7.3"
coverage = {extras = ["toml"], version = "^5.3.1"}
isort = "^5.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
black = {version = "^20.8b1", allow-prereleases = true}
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py38']

[tool.coverage.run]
source = ['./']
